requirements.txt
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
tabulate==0.9.0
openai==1.3.0
gunicorn==21.2.0
